ingestion/requirements.txt
pandas==2.2.0
pyarrow==15.0.0
google-cloud-storage==2.14.0
tqdm==4.66.1
requests==2.31.0